+++ src/utils/parse-text.js
@@ -15,11 +15,11 @@ import { linkHref } from 'social-text-tokenizer/prettifiers';
 import { makeToken, reTokenizer, wordAdjacentChars } from 'social-text-tokenizer/utils';
 import { withCharsAfter, withCharsBefore, withFilters } from 'social-text-tokenizer/filters';
 import { checkboxParser } from './initial-checkbox';
-import { IRAN_FLAG, IRAN_FLAG_EMOJI } from './iran-flag-emoji';
+import { CUSTOM_EMOJI, customEmojiTokenizer } from './emoji-overrides';
 import { SPOILER_END, SPOILER_START, spoilerTags, validateSpoilerTags } from './spoiler-tokens';
 import { isPostLink, isCommentLink } from './post-link-utils';
 
-export { IRAN_FLAG } from './iran-flag-emoji';
+export { CUSTOM_EMOJI } from './emoji-overrides';
 
 const {
   textFormatter: { tldList, foreignMentionServices },
@@ -124,7 +124,7 @@ const shortLinks = withFilters(
   withCharsAfter(wordAdjacentChars.withoutChars('/')),
 );
 
-const iranFlagEmoji = reTokenizer(new RegExp(IRAN_FLAG_EMOJI, 'gu'), makeToken(IRAN_FLAG));
+const customEmojis = customEmojiTokenizer();
 
 export const lineBreaks = reTokenizer(/[^\S\n]*\n\s*/g, (offset, text) => {
   if (text.indexOf('\n') === text.lastIndexOf('\n')) {
@@ -184,7 +184,7 @@ export const parseText = withTexts(
       lineBreaks,
       codeInline,
       codeBlocks,
-      iranFlagEmoji,
+      customEmojis,
     ),
   ),
 );